requirements.txt
pypdf
chromadb
python-dotenv
requests
openai